% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgkmeansr.R
\name{calcWitinSS}
\alias{calcWitinSS}
\title{Calculate the total within cluster sum of squared distance}
\usage{
calcWitinSS(data_cluster, centroid)
}
\arguments{
\item{data_cluster}{Data points in cluster}

\item{centroid}{Centroid of the cluster}
}
\value{
number
}
\description{
Calculate the total within cluster sum of squared distance
}
