% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgkmeansr.R
\name{euc_dist}
\alias{euc_dist}
\title{Euclidean distance between any two points}
\usage{
euc_dist(p1, p2)
}
\arguments{
\item{p1}{First point}

\item{p2}{Second point}
}
\value{
number
}
\description{
Euclidean distance between any two points
}
