% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgkmeansr.R
\name{find_centroid}
\alias{find_centroid}
\title{Find the centroid of a cluster}
\usage{
find_centroid(dat)
}
\arguments{
\item{dat}{Data points}
}
\value{
vector of centroid coordinates
}
\description{
Find the centroid of a cluster
}
