% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgkmeansr.R
\name{init_cent}
\alias{init_cent}
\title{initalize centroids using random or kmpp}
\usage{
init_cent(method, N, K, data)
}
\arguments{
\item{method}{Initialization method}

\item{N}{Number of data points}

\item{K}{Number of clusters}

\item{data}{Data points}
}
\value{
integer Index of data points
}
\description{
initalize centroids using random or kmpp
}
