% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgkmeansr.R
\name{input_preprocessing}
\alias{input_preprocessing}
\title{Convert raw input data to data frame}
\usage{
input_preprocessing(raw, stage)
}
\arguments{
\item{raw}{Raw input}

\item{stage}{`fit` or `predict`}
}
\value{
Cleaned data in data frame format
}
\description{
Convert raw input data to data frame
}
