% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgkmeansr.R
\name{kmpp}
\alias{kmpp}
\title{Find the initial centroids using the kmeans++ approach}
\usage{
kmpp(data, K)
}
\arguments{
\item{K}{The number of clusters}
}
\value{
The indices of the centroids in the input data
}
\description{
Find the initial centroids using the kmeans++ approach
}
\note{
This function is not optimized for performance or efficiency
}
