% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgkmeansr.R
\name{predict}
\alias{predict}
\title{Predict k-means clustering; provide labels for all observations}
\usage{
predict(data, centroids)
}
\arguments{
\item{data}{A to-be-predicted new dataframe with attributes as columns and data points as rows. Data without cluster labels}

\item{centroids}{model from fit function; a data frame of centroids}
}
\value{
Dataframe containing new data and clustering label for each data point
}
\description{
Predict k-means clustering; provide labels for all observations
}
