% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgkmeansr.R
\name{should_stop}
\alias{should_stop}
\title{Check if converged}
\usage{
should_stop(c0, c1, eps)
}
\arguments{
\item{c0}{Previous centroid}

\item{c1}{New centroid}

\item{eps}{Tolerance}
}
\value{
bool
}
\description{
Check if converged
}
