% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgkmeansr.R
\name{kmplot}
\alias{kmplot}
\title{Visualize kmeans results}
\usage{
kmplot(dat)
}
\arguments{
\item{dat}{dataframe. Data with cluster labels}
}
\value{
ggplot object
}
\description{
Visualize kmeans results
}
